resource-manager: add runtime class discovery.
Add a configurable pattern-based matching/mapping of Pod
Spec runtimeHandler values and known runtime classes. It
will be used by subsequent patches to recognise Pods and
Containers of special classes that might require special
care or extra actions when producing or enforcing policy
decisions for resource assignment.

For RunPodSandbox, pick runtimeHandler from the request
and try to map it to a runtime class. For Pod discovery
(IOW, initial sync with the runtime), fire off an extra
PodSandboxStatus request and try to dig out the runtime
handler, class and type details from the implementation-
specific info in the reply.

Having to resort to using implementation-specific data
is unfortunate but seems unavoidable at the moment, as
implementations tend to omit the runtimeHandler passed
to RunPodSandbox from the ListPodSandbox response.